Finally I've managed to complete 1st version of skinned skeleton animation. This is somehow a diffucult path since it's really hard to locate actual problem as I don't have a correct dataset to match. Despite the miserable debugging process, I've learnt a lot from it, both in terms of dual quaternion and skeleton animation data structure.
I think the whole system still requires a lot refactor. The design is somewhat messy, the maths operation is somewhat costy, and the logic is somewhat hard to understand if I read it after a year.
I need to document the whole shit.

However, there's a last loose end that I need to take care, the generation of prev model matrix used for temporal pass. Without it, sophia still behaves like some one living in the phantom space.