I did temporal resolve to shading result which consists both pbr shading and ssr. This worked well before but people found that there were some tiny artifacts, which after some research I found the cause is the low response temporal resolve in static regions of a frame.
The temporal response value "0.98" is for regions that doesn't move both current frame and last frame, doing so I could get a perfect ssr result on very rough surface(Rough surface is extremely noisy for stochastic ssr), while a min-max clipped temporal result(no ghosting) is kicked in if a region is in a movement. To be noticed is that I can't directly use min-max clipped temporal result on the whole frame since it'll keep most of noise fragments and the visual is totally shit.

The artifact occurs inside the border of dynamic/static regions, looking like a tiled-fading temporal resolve, no suprise, I did use tiled neighborhood max motion vector to distinguish whether a region is dynamic/static.

The whole mechanism of this was an attempt to reduce noise of ssr on rough surfaces, but it also brings a lot of artifacts, besides things above, shadows also ghosts seriously(because there's no motion on shadow regions on a static object).

I have no choice but to turn to another solution if these artifacts have to be resolved. This time I separate normal shading result and reflection result. Normal shading result doesn't need low response because it doesn't produce samples across frames, and temporal resolve to it means nothing but temporal anti-alias, using min-max clipped temporal resolve is totally fine. Reflection result on the other hand, consists of skybox reflection and SSR. SSR needs low response due to noise, so previous mechanism will be applied to it and apply to it only.

Now I've got a good balance between ghosting and noise, plus no artifacts(for now). I'm happy:)