Start to do some refactor to texture classes.
Finally I decide to combine these classes in to one, as all their
operations are able to be expressed in a general way, rather than
defining within each of their own class.

Besides, I could put all the factory methods into their base class.

This is a starting step. There'll be a lot of commit relate to this.